T-DAT-901-STG_1/scraper/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-cors==0.8.1
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
chromedriver-autoinstaller==0.6.4
click==7.1.2
colorclass==2.2.2
colorlog==6.9.0
confluent-kafka==2.11.1
croniter==6.0.0
exceptiongroup==1.3.0
faust==1.10.4
frozenlist==1.7.0
h11==0.14.0
idna==3.10
kafka-python==1.4.7
mode==4.3.2
multidict==6.6.4
mypy_extensions==1.1.0
opentracing==1.3.0
outcome==1.3.0.post0
packaging==24.0
propcache==0.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
robinhood-aiokafka==1.1.6
selenium==4.35.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
terminaltables==3.1.10
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
venusian==1.2.0
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.20.1
requests==2.31.0